put cluster functions in LEClusterfunction.h to add some readability
